refactor(browser): separate popup window construction from menus
